hv: fix violations in md.c md.h and md_internal.h for crypto lib
-remove goto
-remove multiple return
-Modify assignment operator in boolean expression
-Modify/fix code style violations
-fix attempt to change parameters passed by value
-fix value need U suffix
-fix use of mixed arithmetic
-fix assigment in expression
-other fixes

Tracked-On: #861
Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/lib/crypto/mbedtls/md_internal.h
@@ -47,7 +47,7 @@ struct mbedtls_md_info_t
     int32_t size;
 
     /** Block length of the digest function in bytes */
-    int32_t block_size;
+    size_t block_size;
 
     /** Digest initialisation function */
     int32_t (*starts_func)( void *ctx );
